+++ tests/no_version/test_unit_tasks_no_version.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
-from fractal_server.tasks.utils import _normalize_package_name
 from fractal_server.tasks.utils import get_absolute_venv_path_v1
+from fractal_server.tasks.v2.utils_package_names import normalize_package_name
 
 
 def test_get_absolute_venv_path(tmp_path, override_settings_factory):
@@ -29,5 +29,5 @@ def test_normalize_package_name():
         "friendly--bard",
         "FrIeNdLy-._.-bArD",
     )
-    outputs = list(map(_normalize_package_name, inputs))
+    outputs = list(map(normalize_package_name, inputs))
     assert len(set(outputs)) == 1