+++ tests/v2/01_schemas/test_task_collection.py
@@ -13,15 +13,42 @@ def test_TaskCollectPipV2():
     Check that leading/trailing whitespace characters were removed
     """
     collection = TaskCollectPipV2(
-        package="  package  ", package_version="  1.2.3  "
+        package="  package  ",
+        package_version="  1.2.3  ",
     )
     assert collection.package == "package"
     assert collection.package_version == "1.2.3"
+
+    collection_none = TaskCollectPipV2(
+        package="pkg", pinned_package_versions=None
+    )
+    assert collection_none.pinned_package_versions is None
+
+    sanitized_keys = TaskCollectPipV2(
+        package="pkg", pinned_package_versions={"    a      ": "1.0.0"}
+    )
+    assert sanitized_keys.pinned_package_versions == dict(a="1.0.0")
 
     with pytest.raises(
         ValidationError, match="Local-package path must be absolute"
     ):
         TaskCollectPipV2(package="not/absolute.whl")
+
+    with pytest.raises(ValidationError):
+        TaskCollectPipV2(
+            package="pkg", pinned_package_versions={";maliciouscmd": "1.0.0"}
+        )
+
+    with pytest.raises(ValidationError):
+        TaskCollectPipV2(
+            package="pkg", pinned_package_versions={"pkg": ";maliciouscmd"}
+        )
+
+    with pytest.raises(ValidationError):
+        TaskCollectPipV2(
+            package="pkg",
+            pinned_package_versions={" a ": "1.0.0", "a": "2.0.0"},
+        )
 
 
 async def test_TaskCollectCustomV2(testdata_path):