+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 **Note**: Numbers like (\#1234) point to closed Pull Requests on the fractal-server repository.
 
+# 2.8.0
+
+* Task group
+    * Add `pinned_package_versions_string` property (\#1980).
+* Task collection
+    * Now both the local and SSH versions of the task collection use the bash templates (\#1980).
+    * Update task-collections database logs incrementally (\#1980).
+    * Support pinned-package versions for SSH task collection (\#1980).
+    * Now `pip install` uses `--no-cache` (\#1980).
+* API
+    * Deprecate the `verbose` query parameter in `GET /api/v2/task/collect/{state_id}/` (\#1980).
+
 # 2.7.1
 
 > WARNING: As of this version, all extras for `pip install` are deprecated and